Refactored route handlers into separate files
